+++ src/components/layout/StatusBar.tsx
@@ -20,7 +20,7 @@ import {
 } from "../../stores/ui-store";
 import { useIntuneStore } from "../../stores/intune-store";
 import { useDsregcmdStore } from "../../workspaces/dsregcmd/dsregcmd-store";
-import { useDeploymentStore } from "../../stores/deployment-store";
+import { useDeploymentStore } from "../../workspaces/deployment/deployment-store";
 import { useSysmonStore } from "../../workspaces/sysmon/sysmon-store";
 import { useEvtxStore } from "../../stores/evtx-store";
